pyproject.toml
[tool.poetry]
name = "lab.ratrig.com"
version = "1.0.0"
description = "Source for labs.ratrig.com"
authors = ["Simon Davie <simon@ratrig.com>"]
license = "CC-BY-NC-SA-4.0"

[tool.poetry.dependencies]
python = "^3.10"
mkdocs-macros-plugin = "^0.5.0"
mkdocs-awesome-pages-plugin = "^2.5.0"
mkdocs-material = "^7.3.0"
Pygments = "^2.10.0"

[tool.poetry.dev-dependencies]
ipython = "*"
ipdb = "*"
black = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
